user-service/requirements.txt
Flask==2.2.5
Flask_SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==2.2.3

